move dependancy checking and running external programs to external.py
